semantics

+++ test/Type_Checking/type_check_one.cl
@@ -11,6 +11,6 @@ type="server";
 #enddb
 void main() {
 
-  int a = 3 + 3.0;
+  float a = 3 + 3.0;
 
 }
